requirements.txt
python-fasthtml
fastlite
httpx
aiofiles
moondream
pillow